[frontend] Rework radar opinions component
