+++ opencti-platform/opencti-front/src/private/components/analyses/opinions/StixCoreObjectOpinions.tsx
@@ -1,7 +1,8 @@
 import React, { FunctionComponent, useCallback, useEffect } from 'react';
-import { useQueryLoader } from 'react-relay';
+import { graphql, PreloadedQuery, usePreloadedQuery, useQueryLoader } from 'react-relay';
+import Typography from '@mui/material/Typography';
 import StixCoreObjectOpinionsRadar, {
-  stixCoreObjectOpinionsRadarFragmentQuery,
+  stixCoreObjectOpinionsRadarDistributionQuery,
 } from './StixCoreObjectOpinionsRadar';
 import useVocabularyCategory from '../../../../utils/hooks/useVocabularyCategory';
 import {
@@ -9,32 +10,64 @@ import {
   StixCoreObjectOpinionsRadarDistributionQuery$variables,
 } from './__generated__/StixCoreObjectOpinionsRadarDistributionQuery.graphql';
 import Loader, { LoaderVariant } from '../../../../components/Loader';
+import { useFormatter } from '../../../../components/i18n';
+import useQueryLoading from '../../../../utils/hooks/useQueryLoading';
+import { StixCoreObjectOpinionsOpenVocabQuery } from './__generated__/StixCoreObjectOpinionsOpenVocabQuery.graphql';
+import StixCoreObjectOpinionsDialog from './StixCoreObjectOpinionsRadarDialog';
 
 interface StixCoreObjectOpinionsProps {
-  stixCoreObjectId: string;
-  variant: string;
-  height: number;
-  marginTop: number;
-  refetch: () => void;
+  stixCoreObjectId: string
+  queryVocabulariesRef: PreloadedQuery<StixCoreObjectOpinionsOpenVocabQuery>
 }
 
-const StixCoreObjectOpinions: FunctionComponent<
+const stixCoreObjectOpinionsOpenVocabQuery = graphql`
+  query StixCoreObjectOpinionsOpenVocabQuery($category: VocabularyCategory!) {
+    vocabularies(category: $category) {
+      edges {
+        node {
+          id
+          name
+          description
+          order
+        }
+      }
+    }
+  }
+`;
+
+const StixCoreObjectOpinionsComponent: FunctionComponent<
 StixCoreObjectOpinionsProps
-> = ({ stixCoreObjectId, variant, height, marginTop }) => {
-  const { typeToCategory } = useVocabularyCategory();
+> = ({
+  stixCoreObjectId,
+  queryVocabulariesRef,
+}) => {
+  const { t } = useFormatter();
+
+  const { vocabularies } = usePreloadedQuery<StixCoreObjectOpinionsOpenVocabQuery>(
+    stixCoreObjectOpinionsOpenVocabQuery,
+    queryVocabulariesRef,
+  );
+  const opinionOptions = vocabularies?.edges
+    .map((edge) => edge.node)
+    .sort((n1, n2) => {
+      if (n1.order === n2.order) {
+        return n1.name.localeCompare(n2.name);
+      }
+      return (n1.order ?? 0) - (n2.order ?? 0);
+    })
+    .map((node, idx) => ({
+      label: node.name.toLowerCase(),
+      value: idx + 1,
+    })) ?? [];
+
   const variables: StixCoreObjectOpinionsRadarDistributionQuery$variables = {
-    // Opininions distribution
     objectId: stixCoreObjectId,
     field: 'opinion',
     operation: 'count',
     limit: 8,
-    // Vocabularies
-    category: typeToCategory('opinion-ov'),
-    // My opinion
-    id: stixCoreObjectId,
   };
   const [queryRef, fetchLoadQuery] = useQueryLoader<StixCoreObjectOpinionsRadarDistributionQuery>(
-    stixCoreObjectOpinionsRadarFragmentQuery,
+    stixCoreObjectOpinionsRadarDistributionQuery,
   );
   const fetchQuery = useCallback(
     () => fetchLoadQuery(variables, { fetchPolicy: 'network-only' }),
@@ -44,27 +77,57 @@ StixCoreObjectOpinionsProps
     () => fetchLoadQuery(variables, { fetchPolicy: 'store-and-network' }),
     [],
   );
+
+  const height = 260;
+
   return (
-    <>
+    <div style={{ height, marginTop: 20 }}>
+      <Typography
+        variant={'h3'}
+        gutterBottom={true}
+        style={{ float: 'left' }}
+      >
+        {t('Distribution of opinions')}
+      </Typography>
+      <StixCoreObjectOpinionsDialog
+        stixCoreObjectId={stixCoreObjectId}
+        opinionOptions={opinionOptions}
+        fetchQuery={fetchQuery}
+      />
+      <div className="clearfix" />
       {queryRef && (
         <React.Suspense
           fallback={
-            <div style={{ height, marginTop }}>
+            <div style={{ height }}>
               <Loader variant={LoaderVariant.inElement} />
             </div>
           }
         >
           <StixCoreObjectOpinionsRadar
-            stixCoreObjectId={stixCoreObjectId}
             queryRef={queryRef}
-            fetchQuery={fetchQuery}
-            variant={variant}
             height={height}
-            marginTop={marginTop}
+            opinionOptions={opinionOptions}
           />
         </React.Suspense>
       )}
-    </>
+    </div>
   );
 };
+
+const StixCoreObjectOpinions: FunctionComponent<Omit<StixCoreObjectOpinionsProps, 'queryVocabulariesRef'>> = (
+  props,
+) => {
+  const { typeToCategory } = useVocabularyCategory();
+  const queryRef = useQueryLoading<StixCoreObjectOpinionsOpenVocabQuery>(stixCoreObjectOpinionsOpenVocabQuery, {
+    category: typeToCategory('opinion-ov'),
+  });
+  return queryRef ? (
+    <React.Suspense fallback={<Loader variant={LoaderVariant.inElement} />}>
+      <StixCoreObjectOpinionsComponent {...props} queryVocabulariesRef={queryRef} />
+    </React.Suspense>
+  ) : (
+    <Loader variant={LoaderVariant.inElement} />
+  );
+};
+
 export default StixCoreObjectOpinions;
